+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Find smallest algorithm
 - Find largest algorithm
 - Selection sort algorithm
+- Factorial function
 
 ## [0.0.1] 2017-10-18